refactor: Configure, InstallIntegration - adds listConnections - adds useProjectId hook - incremental refactor
